add for database jpa function work

+++ aaas-api-user-info/src/main/java/com/shellshellfish/aaas/userinfo/dao/repositories/UserInfoAssetsRepository.java
@@ -1,14 +1,11 @@
 package com.shellshellfish.aaas.userinfo.dao.repositories;
 
 import com.shellshellfish.aaas.userinfo.model.dao.userinfo.UiAsset;
-import com.shellshellfish.aaas.userinfo.model.dao.userinfo.UserAsset;
-import com.shellshellfish.aaas.userinfo.model.dto.user.UserInfoAssectsBrief;
-import com.shellshellfish.aaas.userinfo.model.dto.user.UserInfoBankCards;
-import java.util.List;
+import java.math.BigInteger;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserInfoAssetsRepository extends CrudRepository<UiAsset, Long> {
+public interface UserInfoAssetsRepository extends CrudRepository<UiAsset, BigInteger> {
 
-  UiAsset findByUserId(Long userId);
+  UiAsset findByUserId(BigInteger userId);
 
 }
